fix bug in ready.js

+++ bot/listeners/ready.js
@@ -18,35 +18,37 @@ class ReadyListener extends Listener {
 
     const guildsSettings = await res.json();
 
-    guildsSettings.forEach(({ guild_id, settings }) => {
-      (async () => {
-        const guild = this.client.guilds.get(guild_id);
-        if (guild) {
-          guild.settings = settings;
-        } else if (!guild && this.client.id === '540231971603742761') {
-          const res = await fetch(
-            `${this.client.config.serverHost}/api/guilds/${guild_id}`,
-            {
-              method: 'DELETE',
-              headers: {
-                apikey: process.env.TEST_API_KEY
+    if (guildsSettings.length) {
+      guildsSettings.forEach(({ guild_id, settings }) => {
+        (async () => {
+          const guild = this.client.guilds.get(guild_id);
+          if (guild) {
+            guild.settings = settings;
+          } else if (!guild && this.client.id === '540231971603742761') {
+            await fetch(
+              `${this.client.config.serverHost}/api/guilds/${guild_id}`,
+              {
+                method: 'DELETE',
+                headers: {
+                  apikey: process.env.TEST_API_KEY
+                }
               }
-            }
-          );
+            );
 
-          console.log(`Deleted Guild ${guild_id}`);
-        }
-      })();
-    });
+            console.log(`Deleted Guild ${guild_id}`);
+          }
+        })();
+      });
+    }
 
     const settinglessGuilds = this.client.guilds.filter(
       ({ settings }) => !settings
     );
     if (settinglessGuilds[0]) {
-      settingslessGuilds.forEach(guild => {
+      settinglessGuilds.forEach(guild => {
         (async () => {
           const { settings } = await (await fetch(
-            `${serverHost}/api/guilds/?id=${guild.id}`,
+            `${this.client.config.serverHost}/api/guilds/?id=${guild.id}`,
             {
               method: 'POST',
               headers: {
@@ -57,6 +59,7 @@ class ReadyListener extends Listener {
 
           console.log(`Added Guild ${guild.id}`);
 
+          // eslint-disable-next-line require-atomic-updates
           guild.settings = settings;
         })();
       });
